+++ activesupport/test/configurable_test.rb
@@ -20,13 +20,13 @@ class ConfigurableActiveSupport < ActiveSupport::TestCase
   end
 
   test "adds a configuration hash" do
-    assert_equal({ foo: :bar }, Parent.config)
+    assert_equal({ "foo" => :bar }, Parent.config)
   end
 
   test "adds a configuration hash to a module as well" do
     mixin = Module.new { include ActiveSupport::Configurable }
     mixin.config.foo = :bar
-    assert_equal({ foo: :bar }, mixin.config)
+    assert_equal({ "foo" => :bar }, mixin.config)
   end
 
   test "configuration hash is inheritable" do